Ensure the simple CAB appears below the status bar.
When the contextual action bar is not in overlay mode,
the screen layout is a linear layout.  (screen_simple.xml)

Ensure the standalone CAB appears below the status bar,
consumes the top offset (to avoid content application below it)
and fill in the status area with a guard view for background
protection.

Bug:10014069
Change-Id: I614f16dfa77367a94808aef54710ffebd66e1ca8